Upload step of distrib-nightly-linux.sh should respect the $OHR_SKIP_X86 and $OHR_SKIP_X86_64 env vars
git-svn-id: https://rpg.hamsterrepublic.com/source/wip@12338 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ distrib-nightly-linux.sh
@@ -84,26 +84,29 @@ svn update
 echo "Calling distrib-linux.sh..."
 ./distrib-linux.sh
 
-echo "Uploading 32-bit linux binaries..."
-mv distrib/ohrrpgce-linux-*-wip-x86.tar.bz2 distrib/ohrrpgce-linux-wip-x86.tar.bz2 &&
-scp -p distrib/ohrrpgce-linux-wip-x86.tar.bz2 $UPLOAD_DEST/ohrrpgce/nightly/
-echo "Uploading 64-bit linux binaries..."
-mv distrib/ohrrpgce-linux-*-wip-x86_64.tar.bz2 distrib/ohrrpgce-linux-wip-x86_64.tar.bz2 &&
-scp -p distrib/ohrrpgce-linux-wip-x86_64.tar.bz2 $UPLOAD_DEST/ohrrpgce/nightly/
-rm distrib/ohrrpgce-linux-wip-*.tar.bz2
+if [ -z "$OHR_SKIP_X86" ] ; then
+  echo "Uploading 32-bit linux binaries..."
+  mv distrib/ohrrpgce-linux-*-wip-x86.tar.bz2 distrib/ohrrpgce-linux-wip-x86.tar.bz2 &&
+  scp -p distrib/ohrrpgce-linux-wip-x86.tar.bz2 $UPLOAD_DEST/ohrrpgce/nightly/
+  # The -minimal.zip files are downloaded by the distrib menu.
+  # ohrrpgce-player-linux-bin-minimal.zip is 32-bit
+  echo "Uploading minimal 32-bit ohrrpgce-player binaries..."
+  mv distrib/ohrrpgce-player-linux-bin-minimal-*-wip-x86.zip distrib/ohrrpgce-player-linux-bin-minimal.zip &&
+  scp -p distrib/ohrrpgce-player-linux-bin-minimal.zip $UPLOAD_DEST/ohrrpgce/nightly/
+  rm distrib/ohrrpgce-player-linux-bin-minimal.zip
+fi
 
-# The -minimal.zip files are downloaded by the distrib menu.
-# ohrrpgce-player-linux-bin-minimal.zip is 32-bit
-echo "Uploading minimal 32-bit ohrrpgce-player binaries..."
-mv distrib/ohrrpgce-player-linux-bin-minimal-*-wip-x86.zip distrib/ohrrpgce-player-linux-bin-minimal.zip &&
-scp -p distrib/ohrrpgce-player-linux-bin-minimal.zip $UPLOAD_DEST/ohrrpgce/nightly/
-rm distrib/ohrrpgce-player-linux-bin-minimal.zip
-
-# ohrrpgce-player-linux-bin-minimal-x86_64.zip is 64-bit
-# echo "Uploading minimal 64-bit ohrrpgce-player binaries..."
-mv distrib/ohrrpgce-player-linux-bin-minimal-*-wip-x86_64.zip distrib/ohrrpgce-player-linux-bin-minimal-x86_64.zip &&
-scp -p distrib/ohrrpgce-player-linux-bin-minimal-x86_64.zip $UPLOAD_DEST/ohrrpgce/nightly/
-rm distrib/ohrrpgce-player-linux-bin-minimal-x86_64.zip
+if [ -z "$OHR_SKIP_X86_64" ] ; then
+  echo "Uploading 64-bit linux binaries..."
+  mv distrib/ohrrpgce-linux-*-wip-x86_64.tar.bz2 distrib/ohrrpgce-linux-wip-x86_64.tar.bz2 &&
+  scp -p distrib/ohrrpgce-linux-wip-x86_64.tar.bz2 $UPLOAD_DEST/ohrrpgce/nightly/
+  rm distrib/ohrrpgce-linux-wip-*.tar.bz2
+  # ohrrpgce-player-linux-bin-minimal-x86_64.zip is 64-bit
+  # echo "Uploading minimal 64-bit ohrrpgce-player binaries..."
+  mv distrib/ohrrpgce-player-linux-bin-minimal-*-wip-x86_64.zip distrib/ohrrpgce-player-linux-bin-minimal-x86_64.zip &&
+  scp -p distrib/ohrrpgce-player-linux-bin-minimal-x86_64.zip $UPLOAD_DEST/ohrrpgce/nightly/
+  rm distrib/ohrrpgce-player-linux-bin-minimal-x86_64.zip
+fi
 
 echo "Uploading debian packages..."
 for arch in i386 amd64 ; do
